refactor and added test
