+++ src/entities/BaseEntity.class.ts
@@ -1,0 +1,13 @@
+import { Entity, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn } from 'typeorm'
+
+@Entity()
+export class BaseEntity {
+  @PrimaryGeneratedColumn('uuid')
+  id!: number
+
+  @CreateDateColumn()
+  createdOn!: Date
+
+  @UpdateDateColumn()
+  lastUpdated!: Date
+}
