+++ src/index.ts
@@ -1,13 +1,1 @@
-import { Entity, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn } from 'typeorm'
-
-@Entity()
-export class BaseEntity {
-  @PrimaryGeneratedColumn('uuid')
-  id!: number
-
-  @CreateDateColumn()
-  createdOn!: Date
-
-  @UpdateDateColumn()
-  lastUpdated!: Date
-}
+export * from './entities/BaseEntity.class'
